refactor: `CreateFileAsset`
