+++ Backend/src/Modules/Files/Files.Api/Controllers/Files/FilesResponses.cs
@@ -1,0 +1,5 @@
+﻿using BuildingBlocks.Domain.Files.File.ValueObjects;
+
+namespace Files.Api.Controllers.Files;
+
+public record CreateFileAssetResponse(FileAssetId Id);
