+++ Backend/src/Modules/Files/Files.Api/Extensions/ServiceCollectionExtensions.cs
@@ -1,5 +1,6 @@
 using BuildingBlocks.Api.Converters;
 using BuildingBlocks.Application.Data;
+using Files.Api.Controllers.Files;
 using Files.Application.Data;
 using Files.Application.Extensions;
 using Files.Infrastructure.Data.Extensions;
@@ -23,6 +24,8 @@ public static class ServiceCollectionExtensions
     private static IServiceCollection AddApiServices(this IServiceCollection services)
     {
         TypeAdapterConfig.GlobalSettings.Scan(typeof(FileIdConverter).Assembly);
+
+        services.AddControllers().AddApplicationPart(typeof(FilesController).Assembly);
 
         services.AddScoped<IFilesService, FilesService>();
