+++ Backend/src/Modules/Files/Files.Infrastructure/Data/Configurations/FileAssetConfiguration.cs
@@ -23,10 +23,6 @@ public class FileAssetConfiguration : IEntityTypeConfiguration<FileAsset>
             .IsRequired()
             .HasMaxLength(100);
 
-        builder.Property(n => n.OwnerId)
-            .IsRequired()
-            .HasMaxLength(100);
-
         builder.Property(n => n.Description)
             .IsRequired()
             .HasMaxLength(500);
